requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
distro==1.9.0
fastapi==0.120.1
fastembed==0.7.3
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.9.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
loguru==0.7.3
mmh3==5.2.0
mpmath==1.3.0
numpy==2.3.4
onnxruntime==1.23.2
openai==2.6.1
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
py_rust_stemmers==0.1.5
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
PyYAML==6.0.3
qdrant-client==1.15.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
